Added minimum working example to make a simple query
I added the barebone code needed to make a very simple query on the stackoverflow dataset with bigquery in main.pv. Note that the code requires a json certificate from the google clouds platform to work. That kind of information is rather sensitive, so I  put it in the .gitignore and place the google_credentials_placeholder.json placeholder instead