Polished analysis on tags over time with plot
I polished the analysis of the tag usage over time so that it's faster. I also moved it to the file tag_trends.sql, so that I can use the old distribution_of_first_answer_timequery.sql to work on different questions. Finally. I updated main.py to plot the results of the new query in tag_trends.sql. The labels are not clean yet, but I can work on it.