Updated query to get distribution of waiting times
I updated the query in distribution_of_first_answer_timequery.sql so that it now computes the distribution of waiting time for the accepted answer based on some loose buckets to measure time. I think this concludes the bigquery part I want to showcase, so the next step should involve building a figure and combine this with the older figure into a nice dashboard.