Created dashboard of results
I updated main.py to create a dashboard of the two queries in this project (one for the cumulative number of the most common tags over time, and one for the distribution of time to wait to get the best answer of a question). This also includes some minor changes in the query to compute probabilities instead of raw counts.